duckdb_merge relocated into duckdb utils to future re-use on join operators.
